pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django-routers>=0.2",
    "fastapi>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "slowapi>=0.1.9",
    "uvicorn>=0.35.0",
    "redis>=6.4.0",
    "python-dotenv>=1.1.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.43",
    "opentelemetry-sdk>=1.36.0",
    "opentelemetry-instrumentation-fastapi>=0.57b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.57b0",
    "opentelemetry-instrumentation-requests>=0.57b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.36.0",
    "prometheus-client>=0.22.1",
    "opentelemetry-api>=1.36.0",
    "openai>=1.99.9",
    "bcrypt==4.0.1",
    "pyjwt>=2.10.1",
    "numpy>=2.3.2",
    "matplotlib>=3.10.5",
    "scikit-learn>=1.7.1",
    "email-validator>=2.2.0",
    "boto3>=1.40.21",
    "psutil>=7.0.0",
    "pyyaml>=6.0.2",
    "aiohttp>=3.12.15",
    "ruff>=0.13.2",
    "mypy>=1.18.2",
    "vulture>=2.14",
    "radon>=6.0.1",
]
